Revert "Upgrade Automation on RHEV"
